Refactor pr-check
